requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.4
openpyxl==3.1.2
matplotlib==3.8.2
Pillow==10.2.0
PyMySQL
